pyproject.toml
[project]
name = "orai-paradex-pmm-bot"
version = "0.1.0"
description = "Paradex PMM bot using Backpack SDK"
requires-python = ">=3.10"
authors = [{ name = "Your Name" }]

dependencies = [
    "colorlog==6.9.0",
    "paradex_py==0.4.4",
    "pytest==8.3.4",
    "python-dotenv==1.1.0",
    "backpack-exchange-sdk @ git+https://github.com/Ham1et/backpack_exchange_sdk.git@bugfix/import",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
package-dir = { "orai_paradex_pmm_bot" = "app" }
